% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModSOM.r
\name{ModSOM}
\alias{ModSOM}
\title{Preparation of data for multiSOMe analysis}
\usage{
ModSOM(indata.exp, indata.meth, group.labels, modpar)
}
\arguments{
\item{indata.exp}{gene expression matrix, Should have the same colnames as indata.meth}

\item{indata.meth}{DNA methylation matrix, Should have the same colnames as indata.exp}

\item{modpar}{see manual of modpar function}
}
\value{
modified matrix for input
}
\description{
Preparation of data for multiSOMe analysis
}
